Add float and int
